// workspace/frontend-mdm/src/components/MetadataForms/ElementsList/types.ts
export interface IColumnConfig {
    width: number;
    resizable?: boolean; // false means fixed width
    minWidth?: number;
    maxWidth?: number;
}

export interface IMergedColumnConfig {
    positionIndex?: number;
    sourceFields: string[];
}
export interface IMergedColumns {
    [targetField: string]: IMergedColumnConfig;
}